Round disk percentage in the ListView and make it translatable.
The decimal limitation was lost with the KF5 port.

+++ src/disks.h
@@ -77,7 +77,7 @@ class KDFPRIVATE_EXPORT DiskEntry : public QObject
         QString prettyKBUsed() const { return KIO::convertSizeFromKiB(used); }
         qulonglong kBAvail() const  { return avail; }
         QString prettyKBAvail() const { return KIO::convertSizeFromKiB(avail); }
-        float percentFull() const;
+        int percentFull() const;
         // == comparison
         bool operator==( const DiskEntry & s2 ) const
         {
